requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
lxml==4.2.3
paho-mqtt==1.3.1
requests==2.19.1
urllib3==1.23
